Updated mvp-employee with proper comments

+++ mvp-employee/src/app/employees/employee-list-container/employee-list-container.ts
@@ -1,44 +1,69 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
+
 import { EmployeeService } from '../employee.service';
-import { Router } from '@angular/router';
+import { Observable } from 'rxjs';
+
 
 @Component({
-  selector: 'app-employee-list',
+  selector: 'employee-list',
   templateUrl: './employee-list-container.html',
   styleUrls: ['./employee-list-container.css']
 })
 
 export class EmployeeListContainer implements OnInit{
-    public employeeDetails$:Observable<Employee[]>;
+
+    public employeeDetails$:Observable<Employee[]>;     //Observable for storing employee details
     
-    constructor(private employeeService:EmployeeService,private route:Router) {
-      
-    }
+    constructor(private employeeService:EmployeeService) {}     
 
     ngOnInit()
     {
-        this.getAllEmployee()
+        this.getAllEmployee();
     }
 
-    private getAllEmployee():void
+    /**
+     * Gets all employee details
+     */
+    private getAllEmployee():void     
     {
-        this.employeeDetails$=this.employeeService.getAllData()
+        this.employeeDetails$=this.employeeService.getAllData();
     }
 
+    /**
+     * Deleting Employee with the particular id
+     * @param id  
+     */
     public deleteEmployee(id:number):void
     {
-        this.employeeService.deleteData(id).subscribe(()=>
+        if(confirm("Are you Sure ? "))
         {
-            this.getAllEmployee()
-        })
+            this.employeeService.deleteData(id).subscribe(()=>
+            {
+                alert("Employee Deleted");
+                this.getAllEmployee();
+            })
+        }
+        else
+        {
+            alert("Employee not Deleted");
+        }
     }
+
+    /**
+     * Searching data from Employee List
+     * @param searchInfo 
+     */
     public searchText(searchInfo:string):void
     {
-        this.employeeDetails$=this.employeeService.searchData(searchInfo)
+        this.employeeDetails$=this.employeeService.searchData(searchInfo);
     }
+
+    /**
+     * Sorting data for a particular field
+     * @param sortField 
+     */
     public sortData(sortField:string):void
     {
-        this.employeeDetails$=this.employeeService.sortData(sortField)
+        this.employeeDetails$=this.employeeService.sortData(sortField);
     }
 }
